refactor(react-ui-entities): extract form-component registry into framework
Split the monolithic form-catalog.tsx into one file per component under
src/form-components/ and centralise the EntityFormComponent wrappers that
were scattered across granit-showcase-react:

- phone, url, timezone, email, money — moved from showcase, phone is new
- language — moved from showcase; adds @granit/react-ui-localization peer
- image-upload — moved from showcase; adds @granit/react-ui-blob-storage peer

GRANIT_UI_FORM_COMPONENTS extends STANDARD_FORM_COMPONENTS and is exported
from the package root alongside individual named component exports.
