+++ packages/@granit/react-ui-entities/package.json
@@ -22,7 +22,9 @@
     "@granit/react-localization": "workspace:*",
     "@granit/react-query-engine": "workspace:*",
     "@granit/react-ui": "workspace:*",
+    "@granit/react-ui-blob-storage": "workspace:*",
     "@granit/react-ui-kit": "workspace:*",
+    "@granit/react-ui-localization": "workspace:*",
     "@granit/react-workspaces": "workspace:*",
     "@granit/utils": "workspace:*",
     "@tanstack/react-query": "^5.0.0",
@@ -53,7 +55,9 @@
     "@granit/react-localization": "workspace:*",
     "@granit/react-query-engine": "workspace:*",
     "@granit/react-ui": "workspace:*",
+    "@granit/react-ui-blob-storage": "workspace:*",
     "@granit/react-ui-kit": "workspace:*",
+    "@granit/react-ui-localization": "workspace:*",
     "@granit/react-workspaces": "workspace:*",
     "@granit/types": "workspace:*",
     "@granit/utils": "workspace:*",
