+++ packages/@granit/react-ui-entities/src/form-components/timezone-form-component.tsx
@@ -1,0 +1,19 @@
+import { TimezonePicker } from '@granit/react-ui-kit';
+
+import type { EntityFormComponent } from '@granit/react-entities';
+
+export const TimezoneFormComponent: EntityFormComponent = ({
+  field,
+  value,
+  onChange,
+  readOnly,
+}) => (
+  <TimezonePicker
+    id={`field-${field.propertyName}`}
+    name={field.propertyName}
+    value={typeof value === 'string' ? value : null}
+    onChange={(next) => onChange(next ?? null)}
+    disabled={readOnly}
+    clearable
+  />
+);
